requirements.txt
langchain==0.0.339
numpy
pandas
ipykernel
python-dotenv
google-generativeai
psycopg2
streamlit===1.22.0
tiktoken==0.4.0
protobuf~=3.19.0
langchain_experimental
sentence-transformers
chromadb
ensure
python-box
pyyaml
